+++ packages/core/test/errors.test.js
@@ -2,35 +2,63 @@ const forrest = require('../src/index');
 const { resetState } = require('../src/state');
 
 describe('Errors', () => {
-  let originalLog = null;
-  let fakeLog = null;
-  let logMock = null;
-
   beforeEach(resetState);
-
-  beforeEach(() => {
-    originalLog = console.log;
-    fakeLog = jest.fn();
-    console.log = fakeLog;
-    logMock = () => originalLog(fakeLog.mock.calls.map(($) => $[0]).join('\n'));
-  });
-
-  afterEach(() => {
-    console.log = originalLog;
-  });
 
   describe('getConfig()', () => {
     it('Should provide a meaningful error', async () => {
-      await forrest.run({
-        features: [
-          {
-            target: '$INIT_FEATURE',
-            handler: ({ getConfig }) => {
-              getConfig('foo', '?');
+      const fn = jest.fn();
+      try {
+        await forrest.run({
+          services: [
+            {
+              name: 'foobar',
+              trace: 'abcde',
+              target: '$START',
+              handler: ({ getConfig }) => {
+                getConfig('foo');
+              },
             },
-          },
-        ],
-      });
+          ],
+        });
+      } catch (err) {
+        fn(err);
+      }
+
+      expect(fn).toHaveBeenCalled();
+      expect(fn.mock.calls[0][0].name).toBe('ForrestJSGetConfigError');
+      expect(fn.mock.calls[0][0].message).toContain('getConfig()');
+      expect(fn.mock.calls[0][0].target).toContain('start');
+      expect(fn.mock.calls[0][0].action).toContain('foobar');
+      expect(fn.mock.calls[0][0].trace).toContain('abcde');
+    });
+  });
+
+  describe('getContext()', () => {
+    it('Should provide a meaningful error', async () => {
+      const fn = jest.fn();
+      try {
+        await forrest.run({
+          services: [
+            {
+              name: 'foobar',
+              trace: 'abcde',
+              target: '$START',
+              handler: ({ getContext }) => {
+                getContext('foo');
+              },
+            },
+          ],
+        });
+      } catch (err) {
+        fn(err);
+      }
+
+      expect(fn).toHaveBeenCalled();
+      expect(fn.mock.calls[0][0].name).toBe('ForrestJSGetContextError');
+      expect(fn.mock.calls[0][0].message).toContain('getContext()');
+      expect(fn.mock.calls[0][0].target).toContain('start');
+      expect(fn.mock.calls[0][0].action).toContain('foobar');
+      expect(fn.mock.calls[0][0].trace).toContain('abcde');
     });
   });
 });